fix(webui): extract ConversationItem to module level with React.memo
- Extracted ConversationItem from inline inside ConversationList to module-level ConversationItem.tsx
- Added React.memo via memo(ConversationItemInner) for stable component identity
- Removed unused deleteTarget prop from ConversationItemProps
- Removed stale deleteTarget={deleteTarget} from all 3 render call sites in ConversationList
- Resolved rebase conflict against origin/master
